KUDU-257.  Fix "pure virtual method call" in MaintenanceManager
Fix an issue where MaintenanceManager::FindBestOp was making a pure
virtual function call when it was calling into a MaintenanceOp that was
in the process of being destroyed.

I added MaintenanceOp::Unregister to unregister the op.  Now, the
MaintenanceOp destructor does not unregister the op, but merely checks
to see that the unregistration has been done.

Also, within TabletPeer::Shutdown, call Tablet::UnregisterMaintenanceOps
to avoid maintenance ops triggering while we're in the process of
removing the tablet.

Change-Id: I0f103b0c1cc29f6a361ee76275c08e307f79113a
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2546
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins