client: improve scanner predicate API
The changes here are two-fold:
1. Stop using PB structures in the user-facing API.
2. Allow predicates on keys to be added in any order.

Todd pointed me at ScanSpec and RangePredicateEncoder, and I was able to
reuse them in the client with only minimal refactoring.

For the record, I still don't understand all of the code in
RangePredicateEncoder.EncodeRangePredicates, but I trust Alex and the unit
tests all pass, so...winning?

I'm not thrilled by the bit of ugliness that is CopyPredicateBound, but it
seemed cleaner than giving each DataType a new member that is really only
needed by STRING.

Change-Id: If604008d5ef2669a262cb3e3a624cee753b291f6
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2571
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>