Catalog manager can wait while there's tablets to assign
TestKuduTable failed recently because of this, the catalog manager
is suppose to wake up if there's stuff to do but it doesn't check
that state before starting to sleep. It means that you can create
two tables almost back to back, get the first one created right away
but then the second one takes more than 10 seconds.

Create table stress test is disabled but should at least trigger this
code.

Change-Id: I7fa87a306961ec0a24e7322e233533a9a9c72813
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2555
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins