twitter-demo: fix API consumer to handle large records
Previously, we were assuming that curl gave us chunks of data that were
perfectly aligned with JSON records. This is not in fact the case.  The new
code uses a buffer and splits records on newline.

Tested manually by ingesting the firehose for a minute or so and scanning the
results.

Change-Id: Ifd60fac9ef947b07c61bd9d36673f8cd1e838483
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2561
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins