KUDU-270. Add rw_spinlock for Log::last_entry_op_id_
last_entry_op_id_ is written constantly, but periodically accessed by
various threads including Consensus and Log GC. Add a r/w spinlock
specifically for this field to minimize the possibility of a reader
slowing down writing to the log when this field is accessed.

Also, move the Log locks closer to the fields they guard to improve
cache locality on access.

Change-Id: Ia6d7e71f4d9687c008bdb92f20b07d3f1acab94b
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2573
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>