rw_semaphore: add a debug define which makes the semaphore track the writer stack
This is useful for cases where we accidentally orphan a write lock held on the
semaphore.

We do not currently track the readers, as that is somewhat more complicated:
when a reader unlocks, we don't know which stack it "corresponded" to. The
best we could do is potentially track the N most recent acquirers, but that
is left to future work.

Change-Id: Ie4ad2659507e57cb2b17171000d228063726eda7
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/2585
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins <cloudera-hudson@gerrit.cloudera.org>
Reviewed-by: Michael Percy <mpercy@cloudera.com>