linked_list-test: improve verification to run faster and use less RAM
Instead of using a hashmap, we now collect the data into vectors, sort them,
and use a linear time set intersection algorithm. The big-O notation is actually
worse, since the sort is O(n lg n) but the CPU cache behavior is much better
appending to vectors vs random-accessing a large hashmap. Additionally, the
vector<uint64_t> is more memory efficient because the hashmap uses separate
chaining and thus has linked list pointer costs.

This also adds verification that not too many items are unlinked in the list.
This would have caught the earlier bug where prev_key_ was not updated.
I verified this by commenting out the fix and making sure the test reported an
error.

I also tested the verification path by changing the insert path to drop every
10000th row without calling Apply(). This properly caused a bunch of errors to
print out when verifying.

To test the performance improvement, I ran the test before and after the patch with --seconds_to_run=60. I measured the time to verify as the time between the "Tablet started" message and the "Verified" message.

Before:
  Total test run time: 88.15sec
  Time for verifications: approx 3.3 seconds each

After:
  Total test run time: 79.44sec
  Time for verifications: approx 1.6 seconds each

Change-Id: I04b60562cab5b1facd9db7b588c045a57fcf37f8
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2590
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>