[java client] MapReduce InputFormat
Adds the input format with similar tests to what the output format has.

Removes WritableRowKey and replaces it by using NullWritable. We're not
expecting users to send our Operations or RowResults on the wire and won't
facilitate that.

The way the IO formats are configured is very similar but I'm not sure how
to make it pretty, it seems that any refactoring will just add a lot of dumb
code at the moment.

Change-Id: I18d1c083d315010ea992b45cc241aeeac45b5f82
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2593
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins