Dist. Consensus - Fix race in replica update duplicated op processing
Previously we were going through the operations to find which
ones might be replicated outside the replica lock. That needs
to be done under a lock as it accesses an *Unlocked() method.
Not doing so would make actual repeated messages from the leader
occasionally cause some weird behavior, such as InsertOrDie
into the pending commits map failing on a duplicated key.
This was more likely for the first batch of messages, when
LastReceivedOpId had not been set yet. By placing this block also
under the lock this behavior could not be observed anymore.

In addition to fixing the aforementioned bug this patch adds a DCHECK
to all *Unlocked() methods making sure we've acquired the lock prior
to calling them.

This also injects comm. faults in a couple of other places in the
tests that made this bug more likely (but still not certain) to appear.

Change-Id: I97e955269ad4cbb5f5924a321d1892afa2281698
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2601
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>