linked_list-test: improvements to debuggability of failures
I am working on reproducing and debugging an issue where after running
linked_list-test for many hours I eventually have it fail with 10 or so missing
rows (likely to be one missing write operation). The failure only results in
the count being too low, rather than any missing link in the linked list chain,
which indicates it's probably the last write operation in one of the tablets
going missing.

This patch tries to make that easier to debug in two ways:

* Log the last written chain entries upon completion of insertion. This will
  give us some keys to search for within the stored data.

* Add an option to bootstrap to _not_ delete the replayed logs, even if replay
  was successful. This should help diagnose if in fact the original commit was
  logged but somehow it got lost on replay.

In order to do the latter, I added some new plumbing to external_mini_cluster
to pass arbitrary flags into the TS process.

Change-Id: I684d1c4da2ee2d4039d44153ee99b255cfe43999
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2603
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>