Refactor Leader/Replica *Transactions into Drivers and Transactions
The core goal of this patch is to split transaction coordination (which
is Leader/Replica) from transaction execution (which is mostly the same
on the Leader and Replica sides).

Towards this goals this patch splits *Transactions into the driver classes
(LeaderTransactionDriver and, later on,  ReplicaTransactionDriver) and the actual
transaction implementations (WriteTransaction, AlterSchemaTransaction and
ChangeConfigTransction).

Other notable changes included in this patch:
- Transactions drivers are now thread safe.

Change-Id: I84598ca7aa3b29e06acef9bd6420be5520eefbc3
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2598
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins