[java client] CSV importer MR job
This adds an ImportCsv job that can parse any file with any single byte
separator. The columns can be in a different order than in the schema.
All you need to do is specify the column names on the command line.

Floating points are removed as we currently don't handle them. We assume
numbers in the form of "1.02" and "2.00", meaning that we always have
two decimals, so removing the period is like doing 100x.

Change-Id: I490b2395f8c26499bd23ffa756ce264e2ba68d13
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2619
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins