KUDU-285. Fix TSAN errors in transaction_driver
Many of the functions in this class lock the internal spinlock, and may
then finish the transaction while holding the lock. This could cause
the last refcount of the object to get destroyed, in which case the
object would be destructed while the lock_guard is still in scope. This
patch adds a somewhat ugly hack to increase the refcount on the object
while holding the lock, so that it doesn't go out of scope.

This is somewhat of a bad hack since the refcount increments are extra
atomic instructions which probably aren't necessary. Most likely we can
just unlock the lock_guard before calling tracker->Release, but that is
less straightforward.

I'll leave this JIRA open and repurpose it to do a cleaner fix later.

Change-Id: I095a8831d00f3316506845cfab8922ada95b9e76
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2627
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins