Add a generic resettable heartbeater
In consensus the leader needs to send heartbeat messages to the peers
only if there were no regular messages. This patch adds a heartbeater
that takes a generic function as the heartbeat function and has a Reset()
method that resets the heartbeat interval.

Some of the code is duplicated from the master heartbeater, eventually
we might want to make that one to be based on this one but as there
are currently time contrains that was left as a TODO (along with two
other important aspects, documented in resettable_heartbeater.h, such
as the ability to use a shared ScheduledExecutor and the ability to do
backoff).

Change-Id: I9498c55915b829d31923b07806e8700901a387fe
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2642
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>