Add a simple time based failure detector and monitor
This patch adds a general interface for failure detectors and
a particular implementation for a time based one that simply
triggers a failure callback when is has not received a message
from a node in a certain time.

This also introduces a randomized failure monitor that checks
if any node has been detected as dead every interval where and
interval is mean +- stddev. The point being that we don't want
failures to be detected at the exact same time everywhere if
the failure detectors/monitors are started at the same time.

Change-Id: I33450c78f09c1bf1d2ded596f482fdb8d81c10cf
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4919
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>