[java client] Revamped buffering in KuduSession
The old buffering was just sending the Batch objects as they got full,
possibly overwhelming the tablet servers. This new buffering will send
a new PleaseHoldException (from asynchbase) when we're able to fill a
buffer while another one is outstanding.

Another change is that calling flushTablet now requires passing a Batch,
and the method only flushes if the passed Batch is the one in the buffer.
This is so that the flush timer doesn't try to flush small batches if the
batch it was supposed to flush was already flushed. For example:
 - insert 1 row, new Batch created, flush timer started
 - insert another 999 rows, flush is forced
 - repeat that a couple of times within 1 second (bunch of flush timers created)
 - after 1 second, the 1st flush timer task triggers a flush for whatever is in
   the buffer.
 - subsequently all the other timers will also fire, thus either flushing
   nothing or small batches.

This patch made it possible to insert the 100GB lineitem.tbl dataset
from TPC-H via the ImportCsv job (which also requires changes, that's
coming soon).

Change-Id: I35b4d8ba2abbe41a0fe786eabb1b03cf554e7e60
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2663
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>