Allow to forcefully allocate a new segment/roll the log and to get a Log's FsManager
Upon being elected LEADER, a peer needs to read the Log up to the last written operations.
In order to do this all segments up to the current pont need to be closed and available
as readable segments. We can achieve this by allowing consensus to forcefully roll the Log.

Change-Id: I1ced54707b45884252793dce24f8a08790c1a1d4
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2684
Tested-by: jenkins
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>