Make ChangeConfigTransaction notify (indirectly) TsTabletManager that state is dirty
On consensus role changes TsTabletManager needs to be notified somehow so that it
marks the tablet dirty and eventually reports te new role to master.
This patch addresses this by making TabletPeer receive a function that it can call to
mark the tablet dirty and making it available through a TabletPeer method, allowing
classes with access to TabletPeer to notify TsTabletManager without introducing the
dependency.
Finally this makes ChangeConfigTransaction call the new method on TabletPeer when a
new configuration is flushed into the Tablet's metadata.

Change-Id: Ib1d8e15b2d0a8a222ddd0c749053d114197ebf50
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2683
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>