[java client] Have KuduTableOutputFormat use AUTO_FLUSH_BACKGROUND
It was hard to write a unit test before without fixing KUDU-232, but now that
it's fixed we can enable batching in the output format. It changes the way we
handle errors since they happen in a deferred way, so a counter was added to
be able to quickly identify tasks that have bad rows.

Change-Id: Ie9bad390bdf1294d05899a9b7acef1cdfc698e80
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2695
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>