[java client] Add an option to the input format to specify the projection
Also adds relevant tests, and modifies RowCounter to handle it. Tested on
the TPC-H 100GB data set, makes row counting a lot faster (25 seconds instead
of 1 minute).

A few things were missing to be able to pass an empty Schema to a KuduScanner,
this patch fixes that and adds a unit test that also uses a predicate on a
column that isn't read.

Originally these two changes were meant to be in different patches but it got
really confusing to keep track of, so they are merged here. Sorry!

Change-Id: I4da185961498e3462388baf7aca31bebd5f95a33
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2699
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>