KUDU-297. Fix TSAN error in ReactorThread::DumpRunningRpcs()
The Connection object is writable by only one thread at a time. If we
are in the middle of connection negotiation, do not allow access to the
still-mutable state since there are no locks to protect against data
races. At this time, the only mutable state is the user credentials.

Change-Id: Idd8f2cd34de6c9d090dad0da522f25050dcb3537
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2705
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>