raft: fix issue where writes to FOLLOWER peers would never respond
This adds a new test for a write which goes to a FOLLOWER peer, and
restructures some of the code in the transaction driver so that the
code flow is a little more clear. Now, in all cases, the prepare/replicate
callback will be called twice, even if the prepare fails. This adds
the ability to add some new assertions and makes the "failed to replicate"
case look more like other failure cases.

In the process of changing the code, I also made one of the FutureCallbacks
scope-local instead of being a class member. The variable was only assigned
once, and it actually makes the most sense to allocate it close to where it
is used. This may have a slight performance benefit, and also makes the code
easier since it's obvious these callbacks are "constant" wiring, rather than
something which might change at runtime.

Without the fix applied, the new unit test hangs.

As a follow-up, we need to change the error code in this case to something
which more specifically indicates the reason for the failure. Otherwise,
the client will have to resort to string-matching to know that it should
retry on another member.

Change-Id: Ida1fc0db865858cb3a5b90ab8f2f804dd1689f29
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2706
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins