KUDU-301. Fix constructor-virtual method race in TransactionDriver
TSAN was complaining about a race between adding the TransactionDriver
to the TransactionTracker queue in the constructor, and the Log GC
thread interrogating the tracker and calling a virtual method on that
object before the constructor completed.

Fix that race by utilizing a factory method for TransactionDriver
objects that first constructs and then registers the object with the
TransactionTracker instance via a call to Init().

Change-Id: I3b6a19076e524acf263cb8183d0d295f70ba5658
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2710
Tested-by: Michael Percy <mpercy@cloudera.com>
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>