client: fix an abort race
If Batcher::Abort() is called with an outstanding op, there's a chance that
the op doesn't get properly cleaned up. More specifically, if the op is in
TabletLookupFinished, has checked if the batcher was aborted, but has yet to
change its state, an interleaved Abort() will ignore it. I used sync points
to verify the existence of this race.

To fix this, we need to acquire the batcher lock earlier in
TabletLookupFinished, before checking if the batcher was aborted. That way a
client can't Abort() until after the lock is released, at which point the op
will have been freed (on failure) or changed state (on success).

I think releasing the batcher lock before calling CheckForFinishedFlush()
(in the failure case) is OK, but it's the one thing I'm less sure about.

Change-Id: I0c6865953b0543696a2f4f7c76955c5faff0fee9
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2729
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>