KUDU-306. Create TSAN-annotated alternatives for glog macros
This fixes a TSAN-reported benign data race on a static variable used
in LOG_FIRST_N() by creating an alternative macro called KLOG_FIRST_N().

While we're at it, also add macros for variants of LOG_EVERY_N and
LOG_FIRST_N as well and convert existing code to use those, and attempt
to prevent use of the "stock" macros we are replacing.

Change-Id: I6899e5f74265ba57c5b52eee84b13fb0bae63f3d
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2730
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>