KUDU-308. Fix deadlock between running transactions and TS shutdown
This fixes a deadlock I saw once in looping a test which rapidly started
and stopped a replicated tablet. Because of the replication, the test
exercises the stopping of a TS while ConsensusUpdates are still coming in,
which is what triggered the deadlock:

T1: the TS is trying to shut down and calling TSTabletManager::Shutdown() which
    calls TabletPeer::Shutdown() while holding the tablet manager lock. This
    thread is then waiting for all of the transactions to quiesce.

T2: a ChangeConfig transaction is running, and trying to call TSTabletManager::MarkDirty().
    This blocks on the TSTabletManager lock held by T1.

The solution makes the shutdown method operate on a snapshot of the tablet peer
list so that the lock is free during shutdown.

No new unit test since this was seen by an existing test that I'm adding in the next
commit.

Change-Id: Ib7f61409a47f85a52110a732c7061b3dc3e561b2
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2733
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins