Bug: Double incrementing of configuration sequence numbers
We were incrementing the sequence number of the quorum configuration twice
when running with replication: one in the configuration round of the consensus
algorithm and one in TSTabletManager::OpenTablet().

LinkedListTest was failing because sometimes the leader would lag behind the replicas
on bootstrap and the replicas would see their config seqno increased by TSTabletManager
while the leader wouldn't. In the following bootstrap the leader would try and push
a seqno that was equal to the one on the replicas and the replicas would refuse to execute
the change config transaction. This wasn't a problem before because LocalConsensus
does not increment the config seqno.

This patch solves the problem by removing seqno incrementing from TSTabletMananger as
consensus is really the one that should do it through a configuration round. This
also adds seqno incrementing to LocalConsensus so that the behavior is similar to the
distributed implementation.

This doesn't introduce any specific tests but I can confirm that the linked-list-test
on Todd's branch was failing before this fix and is no longer.

Change-Id: Ia0bb3c7a5e955e3c7faf7b22e7735e3874ad8efd
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2746
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>