[java client] Move the leader's position when followers disconnect
We weren't keeping track of the leader in the list of TabletClients in
RemoteTablet, but we cannot have a unit test right now since we don't have the
required infra in the java side.

Also, thanks to http://gerrit.ent.cloudera.com:8080/#/c/2734/
we now direct all the traffic to the leader, else we retry. That's instead
of sending the traffic to some random follower.

Change-Id: I160ba44bd17713a83bebe0530cd0bf6f2e7d7933
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2757
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins