external_mini_cluster: log executable and flags
This is handy when diagnosing a linked_list-test failure, in particular.
The user can copy the flags from the log into a new terminal to restart
a tablet server or master on the same directory. It also makes it easier
to see in the logs where servers have been restarted.

Change-Id: If408ee664018479f4f1474a84f3cfff27f75530c
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2789
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>