webui: move impala CREATE TABLE statement to master, update it
This moves the CREATE TABLE statement into the table page on the master web UI,
instead of the tablet page on the TS web UI. This makes more sense now that our
client is table-aware.

Also added the master address field which is required by the current version of
Impala integration.

Manually tested on a cluster.

Change-Id: Id13096fc1b1464c4d99aa1795f732fef7423f9cf
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2790
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>