java: add a new class for tablet locations
This avoids exposing the internal classes, and instead gives a "data-only" view
of the tablet locations back to the user. The new class also exposes information
about the replica roles.

I used this successfully in the Impala scan node planner to only send fragments
to the nodes with leaders, and verified that it worked.

Change-Id: I59d3dd2e0a944846946acaecb015528529a0a7ad
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2793
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>