client: allow scanning from non-leader replicas unless otherwise specified
Scanners now try to read the closest replica instead of always reading from the
leader. The closest replica may be inconsistent, and this patch doesn't do
anything to wire in HybridTime. Instead, I used a hammer and added a
SetScanFromLeaderOnly(bool) function which is used to make sure the tests still
pass. In the future we can integrate HybridTime based consistency.

In order to determine which hosts are local, the client now gets the local
hostname and then resolves it at startup. This may not be foolproof, but it's
a first attempt.

This patch should give us more options for locality in Impala.

Change-Id: I5291a7c2519d52a9d5b7b96a2b9d609f93e26839
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2796
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins