Add Multithreaded integration test for consensus
This adds a test that runs multiple clients in different threads against
a quorum and also runs disruptor threads that steal the lock from replicas,
causing timeouts. The point is to emulate situations where the replicas
pause for a long time, such as long synchronization latency when fsync is
off.

At the end of the test it verifies that the leader and the replicas have
the same data.

Change-Id: Ie4644c0c4caf57e173cea82a20f5cd9a428fe892
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2802
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>