KUDU-295. metrics: fix up confusion between gauges and counters
Per Chris Leroy, we should only use Counters for things which monotonically
increase, and should use Gauges for any "measurement in time" values.

This changes the APIs to enforce this - Counters no longer offer the ability
to decrement. Gauges now offer the ability to decrement for cases where we
want to maintain an in-flight count by incrementing and decrementing rather
than by setting values.

I also had to add a static Instantiate method which returns the AtomicGauge<T>
subclass type, since the existing Instantiate methods always returned the
Gauge interface, which is missing the useful methods.

Change-Id: I2cc4b650d0d5d8d07cf72355ad2a3a255951cea3
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2806
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>