Bug Fix: Don't use the commit callback in LEADER transactions if the Apply() failed
When the Apply() fails we don't use the commit callback since the client reply is handled
inline. However we were creating an OperationStatusTracker for this commit with the ctor
that does accept a callback and passing a NULL one. Hence the status tracker would try to
use it and SIGSEGV. This makes it so that we use the appropriate ctor in this situation.

Change-Id: I1e10d0dcb2894d2a2c2bc5327b21b745c63b96e0
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2837
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins