Bug Fix: Avoid self-deadlock when the leader aborted, in replica transactions
When replica transactions receive an OP_ABORT from the leader they would correcly
call HandlePrepareOrLeaderCommitFailure() but would do so in the same thread as the
original RaftConsensus::Update() call. This would deadlock as the replica would try
to call RaftConsensus::Commit() while holding the lock from the RaftConsensus::Update()
call.

This patch addresses this by making it so that OP_ABORT commits are handled the same
way as regular commits, i.e. by a different thread on the 'apply_executor_'.

Change-Id: I8f9065981174ba1b913eece449e58661225e4cec
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2838
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins