linked_list-test: run with replication
This changes the linked list integration test to run with replication 3.
It also required a change to bootstrap to allow out-of-order op ids in
COMMIT messages. David hand-waved a bit about why this is OK, but I wasn't
completely convinced, so I left in a WARNING message when we hit these,
so we don't forget to come back to it (KUDU-333).

The test also had to be changed so that the servers use the same RPC port
after restarting. This is due to the fact that consensus doesn't properly
refresh quorum information from the master when the participants change
RPC addresses (KUDU-334)

Change-Id: Ic87310c8ff08856a456beb3818d9157baf6e3832
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2833
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>