mvcc: clean up "wait for snapshot" code
This cleans up some of the methods around waiting for a particular snapshot to
become "committed". The methods were a bit confusing -- an MvccSnapshot is
already a representation of which transactions are committed, so it doesn't
make much sense to "wait for the transactions in a snapshot to become
committed".

The new methods take a timestamp, and produce an MvccSnapshot at that
timestamp, with the guarantee that any transactions having earlier timestamps
are already committed. These methods are meant to be called on timestamps in
the past, and thus they also guarantee that no new transactions may be started
which have a timestamp lower than the passed-in timestamp.

In addition to making more sense, this also fixes a likely bug with snapshot
scans. Previously, we would wait for all transactions lower than the given
timestamp to be committed, but then we would pass the original MvccSnapshot
into the scanner. The MVCC code would then ignore any of those transactions
which were now committed, even though the whole point of waiting for them was
so that the client would read them.

Change-Id: I7196a9b19a91749087a446086f7741fe22a72a34
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2840
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins