Don't hold component_lock for the lifetime of transactions
Prior to this commit, a write transaction held Tablet::component_lock_
from its PREPARE phase all the way through to COMMIT. This could cause
a messy deadlock:

- a replica transaction is in between PREPARE and COMMIT, holding component_lock
  in R mode
- a background Flush thread comes along and tries to take component_lock in W mode
- the replica prepare executor tries to start a new transaction, which blocks
  trying to take component_lock in R mode (since a writer is pending, no new
  readers can acquire the lock)
- because the prepare executor is now blocked, the leader stops getting ACKs
  and never sends the COMMIT for the original transaction, resulting in
  a deadlock

The new design restores component_lock_ to its original purpose: locking only
the current "components" of a tablet (ie its constituent rowsets). The lock
is treated like a proper spinlock and only held in tight scopes when grabbing
a reference to the current rowsets.

This opens up some potential issues, since we need to be able to use MVCC
to determine whether a given transaction was written into the DuplicatingRowSet
during a compaction. If there is no correspondence between MVCC transaction IDs
and the components that a transaction writes to, that would be impossible.

To combat this issue, WriteTransactions now capture a reference to the current
Components during their Prepare phase, and grab an MVCC transaction ID under
the same lock. This allows compactions to perform a bit of a clever dance in
order to determine which transactions saw the "old" components (ie without the
duplicating rowset) and which saw the "new" ones (ie with the duplication).
This dance is documented in Tablet::DoCompactionOrFlush.

We also need to ensure during an MRS Flush that we don't start writing out
the old MRS before any transactions that would write into it are committed.
We do this using a new MVCC function which waits for all in-flights to commit,
after we swap in the new MRS. Because transactions take the Component reference
at the same time as they take an MVCC ID, this "quiesce" phase ensures that no
more transactions point to the old Components when we start flushing.

Tested an earlier version of this patch on a cluster doing some big imports.
They used to deadlock but don't anymore.

This almost certainly breaks thread-safety of ALTER TABLE, which used to rely
on component_lock_ as a sort of tablet-wide catch-all lock. We need a stress
test for that which shows the breakage, since right now tests aren't failing
despite what seems like a clear issue. Given that this patch is already somewhat
complex, I want to leave that for a later commit (KUDU-382).

Also left as a follow-on is changing component_lock_ to an rw_spinlock, given
that it is no longer held across thread switches.

Change-Id: I35b1ed6cc8cb1408ef8f7d8606c3bf84d95be950
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3115
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>