client: try other replicas during new scan failures
This patch introduces very basic replica failover into the C++ client. How
basic?
1. Only during scans, not inserts.
2. Only when starting a new scan, not continuing an existing one.
3. Only when a TS is unreachable, not when it returns "tablet not found".

When a replica fails, we mark it as such instead of removing it outright.
This is partly because the MetaCache contract forbids deleting TSes, and
partly because it simplifies the handling of "forgiving" failed replicas,
should we choose to implement that in the future. Failed replicas are
ignored during replica selection.

The linked list test change is straight-forward, though it doesn't guarantee
the workings of the patch with every run. To do that, I wrote a client unit
test that uses a new replica selection policy and MetaCache internals to
verify failover.

I also fixed two ancillary issues here:
1. MiniCluster wasn't managing its count of tablet servers properly when new
   ones were added.
2. ClientTest.TestScanMultiTablet was using the wrong table for inserts and
   scans. AFAICT there's no way of verifying this without exposing many
   client internals to the test.

Change-Id: I02077f8b7a2dc4557c330bb40ad8be71334dd97d
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2868
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins