trace: fix microsecond delta computation to not go negative
Previously was computing the delta between successive trace entries by
comparing only the microsecond portion of the clock. So, if one entry was at
23.9 seconds, and the next was at 24.1, we'd see that as a delta of -0.8s
instead of +0.2s.

No test for this because making a reliable test would be quite difficult.  I
verified manually by running linked_list-test with a concurrent 'dd' to the
same disk to make some traces print out:

  0605 11:47:22.168553 (+    13us) log.cc:346] Serialized 522 byte log entry
  0605 11:47:22.168575 (+    22us) raft_consensus.cc:462] Waiting on the replicates to finish logging
  0605 11:47:23.538566 (+1369991us) raft_consensus.cc:464] finished
  0605 11:47:23.538569 (+     3us) raft_consensus.cc:376] Updating watermarks
  0605 11:47:23.538581 (+    12us) inbound_call.cc:81] Queueing success response

Prior to the fix, the third trace message would have shown +369991us (0.4s)
instead of the proper +1369991us (1.4s).

Change-Id: I39ab6e78ad08785718dd3024d0ca4beb7d12dd34
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2865
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>