KUDU-319. Time out hung unit tests more gracefully
When we compress test output, often we save little or no output when
cmake sends a kill at 900 seconds because buffers are not flushed.

Instead of using cmake to kill excessively long-lived tests, use a timer
initiated in test_main.cc that does a graceful kill of the process via
LOG(FATAL).

Change-Id: I5c4123886e502ea546b3db2a3934b872304161fc
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2923
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>