Created WriteOperation abstraction
Abstracted WriteOperation to generalize for the insert/update/delete operations.
Refactored Insert calls to WriteOperation calls in all appropriate locations.

Pulled the entire WriteOperation hierarchy out to a separate file.

In order to simplify the client's API for the KuduSession class, created
an Insert version of Apply(...) so that the client could
call Apply on pointers to the derived class of WriteOperation.

Delete/Update classes have not been added yet, so this does not resolve KUDU-264
completely.

Change-Id: I03f761595da0a1959635b9459c5aece79bdf193a
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2880
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>