client: when TS unreachable, mark all hosted replicas as failed
The original patch that introduced replica failover behavior did so
conservatively: when a scan fails, it marks that combination of replica
and TS as failed.

We can be more aggressive though; we should mark all replicas hosted by the
TS as failed, not just ones belonging to the current tablet.

Change-Id: I2fe3e8f79008dbde169b906b55243dd74cd16f1f
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2926
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>