client: prevent master lookup storms
It's a feature that the Java client has had for ages, and I've copied it
shamelessly. Bonus: we can get rid of the periodic flushing that some
client tests perform to avoid overwhelming the master.

I'm not sure if a sleeping lock is most appropriate, but I don't see us
using boost TimedLockables anywhere, so POSIX semaphores it is.

Change-Id: I7fed0a0aa96458e3227bac8ca2b095e85b27de0d
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2896
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>