linked_list-test: fix failure when verifying from lagging replica
After we load the data, we immediately perform a scan to see that it
is all there. However, it's possible the scan would read from a lagging
non-LEADER replica. This patch changes the behavior to keep trying
to verify for a bit to give a chance for the lagging replica to catch up.

Change-Id: I50359aeb504072b4d173f4d8c4bb661ff698529f
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2910
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins